fix(core): resolve type cast errors and API timeouts
- Add 10-second timeout to Yahoo Finance HTTP requests to prevent silent hangs
- Fix indicator engine crash by explicitly using 0.0 in ternary operators
- Safely parse Yahoo Finance JSON arrays to double to prevent int cast errors
- Add verbose error logging to ScreenerService

+++ lib/domain/indicator_engine.dart
@@ -881,9 +881,9 @@ BXtrenderResult calcBXtrender(
         ? 100 -
               100 / (1 + upDiff[i] / (downDiff[i] != 0 ? downDiff[i] : 0.001)) -
               50
-        : 0,
+        : 0.0,
   );
-  final maShort = t3(shortXR.cast<double>(), 5);
+  final maShort = t3(shortXR, 5);
 
   final isLong = List<bool>.filled(n, false);
   final isShort = List<bool>.filled(n, false);
